functions/mortgage_rate/requirements.txt
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
idna==2.9
jmespath==0.9.5
numpy==1.18.2
pandas==0.24.2
psycopg2-binary==2.8.5
pycrypto==2.6.1
pydash==4.7.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
xlrd==1.2.0
xmltodict==0.12.0
